requirements.txt
langchain 
langchain-community 
langchain-chroma 
langchain-openrouter 
langchain_huggingface
sentence-transformers 
python-dotenv